+++ test/functional/endpoints.spec.ts
@@ -103,6 +103,7 @@ describe('Method Unit Tests', () => {
                 { name: 'getReportAsCSV', stub: 'get', options: {reportId: 123}, expectedRequest: {url: "reports/123", accept: constants.acceptHeaders.textCsv }},
                 { name: 'getReportPublishStatus', stub: 'get', options: {reportId: 123}, expectedRequest: {url: "reports/123/publish" }},
                 { name: 'setReportPublishStatus', stub: 'put', options: {reportId: 123}, expectedRequest: {url: "reports/123/publish" }},
+                { name: 'deleteReport', stub: 'delete', options: {reportId:123}, expectedRequest: {url: "reports/123"}},
             ]
         },
         {
@@ -196,11 +197,11 @@ describe('Method Unit Tests', () => {
                 { name: 'addDiscussionComment', stub: 'post', options: {sheetId:123, discussionId: 234}, expectedRequest: {url: "sheets/123/discussions/234/comments"}},
                 { name: 'deleteDiscussion', stub: 'delete', options: {sheetId:123, discussionId: 234}, expectedRequest: {url: "sheets/123/discussions/234"}},
                 // get
-                { name: 'getSheet', stub: 'get', options: {}, expectedRequest: {url: "sheets"}},
+                { name: 'getSheet', stub: 'get', options: {sheetId: 123}, expectedRequest: {url: "sheets/123"}},
                 { name: 'listSheets', stub: 'get', options: undefined, expectedRequest: {url: "sheets"}},
-                { name: 'getSheetAsCSV', stub: 'get', options: {}, expectedRequest: {url: "sheets", accept: constants.acceptHeaders.textCsv}},
-                { name: 'getSheetAsExcel', stub: 'get', options: {}, expectedRequest: {url: "sheets", accept: constants.acceptHeaders.vndMsExcel, encoding:null}},
-                { name: 'getSheetAsPDF', stub: 'get', options: {}, expectedRequest: {url: "sheets", accept: constants.acceptHeaders.applicationPdf, encoding:null}},
+                { name: 'getSheetAsCSV', stub: 'get', options: {sheetId: 123}, expectedRequest: {url: "sheets/123", accept: constants.acceptHeaders.textCsv}},
+                { name: 'getSheetAsExcel', stub: 'get', options: {sheetId: 123}, expectedRequest: {url: "sheets/123", accept: constants.acceptHeaders.vndMsExcel, encoding:null}},
+                { name: 'getSheetAsPDF', stub: 'get', options: {sheetId: 123}, expectedRequest: {url: "sheets/123", accept: constants.acceptHeaders.applicationPdf, encoding:null}},
                 { name: 'getSheetVersion', stub: 'get', options: {sheetId: 123}, expectedRequest: {url: "sheets/123/version"}},
                 { name: 'listOrganizationSheets', stub: 'get', options: undefined, expectedRequest: {url: "users/sheets"}},
                 // summaries