+++ test/functional/sheets.spec.ts
@@ -32,16 +32,16 @@ describe('Client Unit Tests', () => {
       'should not change base URL when getSheetVersion is called first',
       async () => {
         // First call to getSheet
-        await smartsheet.sheets.getSheet({ id: 100 });
-        expect(spyGet.mock.calls[0][0]).toHaveProperty('url', 'sheets');
+        await smartsheet.sheets.getSheet({ sheetId: 100 });
+        expect(spyGet.mock.calls[0][0]).toHaveProperty('url', 'sheets/100');
 
         // First call to getSheetVersion
         await smartsheet.sheets.getSheetVersion({ sheetId: 100 });
         expect(spyGet.mock.calls[1][0]).toHaveProperty('url', 'sheets/100/version');
 
         // Second call to getSheet
-        await smartsheet.sheets.getSheet({ id: 100 });
-        expect(spyGet.mock.calls[2][0]).toHaveProperty('url', 'sheets');
+        await smartsheet.sheets.getSheet({ sheetId: 100 });
+        expect(spyGet.mock.calls[2][0]).toHaveProperty('url', 'sheets/100');
       }
     );
 
@@ -49,16 +49,16 @@ describe('Client Unit Tests', () => {
       'should not change base URL when getOrganizationSheets is called first',
       async () => {
         // First call to getSheet
-        await smartsheet.sheets.getSheet({ id: 100 });
-        expect(spyGet.mock.calls[0][0]).toHaveProperty('url', 'sheets');
+        await smartsheet.sheets.getSheet({ sheetId: 100 });
+        expect(spyGet.mock.calls[0][0]).toHaveProperty('url', 'sheets/100');
 
         // First call to getSheetVersion
         await smartsheet.sheets.listOrganizationSheets();
         expect(spyGet.mock.calls[1][0]).toHaveProperty('url', 'users/sheets');
 
         // Second call to getSheet
-        await smartsheet.sheets.getSheet({ id: 100 });
-        expect(spyGet.mock.calls[2][0]).toHaveProperty('url', 'sheets');
+        await smartsheet.sheets.getSheet({ sheetId: 100 });
+        expect(spyGet.mock.calls[2][0]).toHaveProperty('url', 'sheets/100');
       }
     );
   });