+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 - Added support for PATCH /2.0/reports/{reportId}/definition endpoint
 - WireMock integration tests for contract testing for PATCH /2.0/reports/{reportId}/definition endpoint
+- Support for POST /reports/{reportId}/scope endpoint
+- Support for DELETE /reports/{reportId}/scope endpoint
+- Support for DELETE /reports/{reportId} endpoint
+
+## [4.7.2] - 2026-03-17
+### Fixed
+- App crashing after updating to v4.7.1 with logLevel set to 'info' [#158](https://github.com/smartsheet/smartsheet-javascript-sdk/issues/158)
+- getSheet API returns 404 Not Found for existing sheet [#161](https://github.com/smartsheet/smartsheet-javascript-sdk/issues/161)
 
 ## [4.7.1] - 2026-02-12
 ### Added